requirements.txt
google-genai==1.12.1
python-dotenv==1.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
edge-tts==7.0.2
frozenlist==1.7.0
gTTS==2.5.4
idna==3.10
multidict==6.4.4
pocketsphinx==5.0.4
propcache==0.3.2
PyAudio==0.2.14
pycparser==2.22
requests==2.32.4
sounddevice==0.5.2
SpeechRecognition==3.14.3
srt==3.5.3
tabulate==0.9.0
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.1
